Modularize the Mongoose and SQL into adapters.
Create a nice interface for creating custom adapters while lazily load the default adapters (Mongoose and SQL).
